old-tests/submission/MOSQUITO/event-source/line-starts-with-colon.html

<!DOCTYPE html>
<html>
<head>
	<title>Lines must be processed, in the order they are received, as follows: If the line starts with a U+003A COLON character, ignore the line:</title>
  <meta charset="utf-8" />
  	<meta name=viewport content="width=device-width, maximum-scale=1.0, user-scalable=no" />
	  <link rel="author" title="Fabrice Clari" href="mailto:f.clari@inno-group.com">
	  <link rel="author" title="Dimitri Bocquet" href="mailto:Dimitri.Bocquet@mosquito-fp7.eu">
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="/resources/WebIDLParser.js"></script>
	<script src="/resources/idlharness.js"></script>
</head>
<body>
	<h1>Lines must be processed, in the order they are received, as follows: If the line starts with a U+003A COLON character, ignore the line:</h1>
	<div id="log"></div>
  <script>
    var s = new EventSource('http://html5.mosquito-fp7.eu/messages/t0071');
    t = async_test("If the line starts with a U+003A COLON character (:), ignore the line.");

    s.onopen = function(e){ 
    	/* Connection opened. */
    };

    s.onmessage = function(e){
    	var dat = e.data.split("\n");
    	console.log(dat);
    	t.step(function(){assert_equals(dat[1], undefined)});
    	t.done();
    };
    	
    s.onerror = function(e){
    	/* Connection closed. */
    };

  </script>
  </div>
</body>
</html>
